Revert 245816ab2721f386184878d71be01143d207abf5 (fixes #348)
Since 245816ab2721f386184878d71be01143d207abf5, IntersectionObserverArticle
assumes that its children do not change unless the number of children changes.

This is not the case with the notification overlay, which resulted in the
checkmark of notification cleaning mode not updating unless scrolling to make
notifications appear/disappear.

This change may negatively impact performances.