Simplify scrolling code by postponing scroll restore to after content is drawn
One reason we had to use such a hack is because statuses are render()-ed
multiple time in the pipeline when they are auto-collapsed, which meant the
ScrollableList lifecycle functions were called before statuses had their final
height. By postponing the scroll repositioning to after the content is drawn,
we can bypass such limitations while simplifying the code.