Use Sanitize to strip invalid codelang attributes
Instead of passing arbitrary user input to the frontend and checking if it's valid there, use Sanitize to strip invalid code languages.

For that a yaml file was added containing all valid language codes, which needs to be kept in sync with highlight.js.
This is inconvenient, but shouldn't be a problem.

Signed-off-by: Plastikmensch <plastikmensch@users.noreply.github.com>