Remove invalid `data-codelang` in backend
Adds a new yml file containing all valid code languages which are supported by highlight.js

This file has been created by a script as the list of supported languages is unreliable.

This is not ideal, but highlight.js made it clear they won't add new languages natively. Instead additional language support is provided by third-party packages.
If such a third-party package is installed, this file needs to be updated to support the new aliases.

When an unsupported language is given, the `data-codelang` attribute is removed using a new transformer.

Signed-off-by: Plastikmensch <plastikmensch@users.noreply.github.com>